requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
python-dotenv==1.0.0
email-validator==2.1.0.post1
pytest==7.4.3
pytest-cov==6.0.0
pytest-flask==1.3.0
flask-swagger-ui==4.11.1
gunicorn==21.2.0
black==23.11.0
flake8==6.1.0
